+++ apps/sv/src/main/scala/com/daml/network/sv/util/CnsUtil.scala
@@ -9,7 +9,7 @@ import scala.util.{Failure, Success, Try}
 object CnsUtil {
 
   val entryNameSuffix = ".unverified.cns"
-  val entryNameRegex = s"[a-z0-9_-]{1,${40 - entryNameSuffix.length()}}\\.unverified\\.cns"
+  val entryNameRegex = s"[a-z0-9_-]{1,${60 - entryNameSuffix.length()}}\\.unverified\\.cns"
   val entryUrlLength = 255
   val validEntryUrlSchemes = Seq("http", "https")
   val entryDescriptionLength = 140